cephadm: move ContainerEngine to container_engine_base.py
Pair-programmed-with: Adam King <adking@redhat.com>
